Separate Python 3 worker
